naPO/manualAdd-on/requirements.txt
playwright>=1.40.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
lxml>=5.0.0
